refactor(ManageInstance): reuse ManageEntityHeader
