+++ src/hooks/common/useExecutableActions.ts
@@ -47,6 +47,7 @@ export type UseExecutableActionsReturn = {
   stopDisabled: boolean
   startDisabled: boolean
   rebootDisabled: boolean
+  deleteDisabled: boolean
   logsDisabled: boolean
   streamDetails?: StreamPaymentDetails
   handleStop: () => void
@@ -194,6 +195,7 @@ export function useExecutableActions({
   const stopDisabled = !isPAYG || !isAllocated || !crn
   const startDisabled = !isPAYG || isAllocated || !crn
   const rebootDisabled = !isAllocated || !crn
+  const deleteDisabled = !executable
 
   const handleStop = useCallback(
     () => handleSendOperation('stop'),
@@ -323,6 +325,7 @@ export function useExecutableActions({
     stopDisabled,
     startDisabled,
     rebootDisabled,
+    deleteDisabled,
     logsDisabled: !logs,
     handleStop,
     handleStart,
